% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrad.R
\name{corrad}
\alias{corrad}
\title{Correlation anomaly detection}
\usage{
corrad(DA, DB, rho, lag = FALSE, eliminate_phase = FALSE)
}
\arguments{
\item{DA}{reference data matrix.}

\item{DB}{test data matrix.}

\item{rho}{penalty weight for graphical lasso.}

\item{lag}{boolean. defalut is FALSE.
if TRUE, it corrects time lag between time series
by maximizing cross-correlation.}

\item{eliminate_phase}{boolean. defalut is FALSE.
if TRUE, it executes foulier-amplitude-based CAD (CAD-II).}
}
\description{
Correlation anomaly detection
}
