% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_corrad.R
\name{ex_corrad}
\alias{ex_corrad}
\title{Extended correlation anomaly detection (ECAD-I & ECAD-II)}
\usage{
ex_corrad(
  DA,
  DB,
  rho,
  dlt,
  eliminate_phase = FALSE,
  smooth = FALSE,
  bandwidth = 0.2
)
}
\arguments{
\item{DA}{reference data matrix.}

\item{DB}{test data matrix.}

\item{rho}{penalty weight for graphical lasso.}

\item{dlt}{time interval of recorded data.}

\item{eliminate_phase}{boolean. defalut is FALSE.
if TRUE, it executes foulier-amplitude-based ECAD (ECAD-II).}

\item{smooth}{boolean. if TRUE, apply parzen-window smoothing to DA & DB.}

\item{bandwidth}{if smooth == TRUE, it specifies bandwidth of parzen-window.}
}
\description{
Extended correlation anomaly detection (ECAD-I & ECAD-II)
}
