% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parzen.R
\name{parzen_window}
\alias{parzen_window}
\title{Parzen window smoothing}
\usage{
parzen_window(freq, u)
}
\arguments{
\item{freq}{frequency}

\item{u}{the parameter of Parzen window}

\item{power}{power spectral densities to which you want to apply smoothing}

\item{bandwidth}{the bandwidth of Parzen window}
}
\description{
Parzen window smoothing
}
