% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{ROC}
\alias{ROC}
\title{Parzen window smoothing}
\usage{
ROC(score, inx_anomaly)
}
\arguments{
\item{score}{anomaly score for which roc is computed}

\item{inx_anomaly}{indecies of abnormal variable}
}
\description{
Parzen window smoothing
}
